git-svn-id: https://5-10/svn/mxsrclib/trunk@236 63356120-0428-6b47-bcd6-cd59724df219

+++ irssysutils.h
@@ -20,12 +20,27 @@
 #undef max
 
 namespace irs {
+
 // Вид числа
-enum { num_precision_default = static_cast<int>(-1) };
-enum num_mode_t{num_mode_invalid, num_mode_general, num_mode_fixed, num_mode_scientific,
-  num_mode_default = num_mode_general};
-enum num_base_t {num_base_invalid, num_base_dec, num_base_hex, num_base_oct,
-  num_base_default = num_base_dec};
+enum {
+  num_precision_default = -1
+};
+
+enum num_mode_t {
+  num_mode_invalid,
+  num_mode_general,
+  num_mode_fixed,
+  num_mode_scientific,
+  num_mode_default = num_mode_general
+};
+
+enum num_base_t {
+  num_base_invalid,
+  num_base_dec,
+  num_base_hex,
+  num_base_oct,
+  num_base_default = num_base_dec
+};
 
 // Универсальная функция перевода чисел в текст
 template<class T>
